API

Qecsim

Qecsim — Module

Package for simulating quantum error correction using stabilizer codes.

Qecsim.doubler — Method

doubler(x)

Return 2x.

Examples

julia> doubler(3)
6

PauliTools

Qecsim.PauliTools — Module

Tools for Pauli strings and binary symplectic vectors / matrices.

Qecsim.PauliTools.bsp — Method

bsp(A::BitArray, B::BitArray)

Return the binary symplectic product of A with B, given in binary symplectic form.

The binary symplectic product $\odot$ is defined as $A \odot B \equiv A \Lambda B \bmod 2$ where $\Lambda = \left[\begin{smallmatrix} 0 & I \\ I & 0 \end{smallmatrix}\right]$.

Examples

julia> using Qecsim: PauliTools as PT

julia> a = BitVector([1, 0, 0, 0]);  # XI

julia> b = BitVector([0, 0, 1, 0]);  # ZI

julia> PT.bsp(a', b)
1

julia> using Qecsim: PauliTools as PT

julia> stabilizers = BitMatrix(  # 5-qubit stabilizers
       [1 0 0 1 0 0 1 1 0 0     # XZZXI
        0 1 0 0 1 0 0 1 1 0     # IXZZX
        1 0 1 0 0 0 0 0 1 1     # XIXZZ
        0 1 0 1 0 1 0 0 0 1]);  # ZXIXZ

julia> error = BitVector([0, 0, 1, 1, 0, 0, 1, 0, 1, 0]);  # IZXYI

julia> PT.bsp(stabilizers, error)
4-element Vector{Int64}:
 0
 1
 1
 0

Qecsim.PauliTools.pauli_to_bsf — Method

pauli_to_bsf(pauli)

Convert the Pauli operator(s) to binary symplectic form. A single string is converted to a vector. A collection of strings is converted to a matrix where each row corresponds to a pauli.

Examples

julia> using Qecsim: PauliTools as PT

julia> PT.pauli_to_bsf("XIZIY")
10-element BitVector:
 1
 0
 0
 0
 1
 0
 0
 1
 0
 1

julia> using Qecsim: PauliTools as PT

julia> PT.pauli_to_bsf(["XIZIY", "IXZYI"])
2×10 BitMatrix:
 1  0  0  0  1  0  0  1  0  1
 0  1  0  1  0  0  0  1  1  0