Models

BasicModels

Qecsim.BasicModels — Module

Basic stabilizer codes

Qecsim.BasicModels.BasicCode — Type

BasicCode <: StabilizerCode

Basic code defined by its stabilizers and logical operators.

Qecsim.BasicModels.BasicCode — Method

BasicCode(pauli_stabilizers, pauli_logical_xs, pauli_logical_zs;
          nkd=nothing, label=nothing)

Construct a basic code from string representations of stabilizers and logical operators.

Paulis are expressed as strings of capitalized I, X, Y, Z characters, with one character per physical qubit. Logical X and Z operators are in matching order, with one of each for each logical qubit. Optional nkd defaults to n and k evaluated and d nothing. Optional label defaults to "Basic [n, k, d]".

Examples

julia> using Qecsim.BasicModels

julia> code = BasicCode(["ZZI", "IZZ"], ["XXX"], ["IIZ"]);  # 3-qubit repetition

julia> validate(code)  # no error indicates operators satisfy commutation relations

julia> nkd(code)  # default nkd
(3, 1, nothing)

julia> label(code)  # default label
"Basic [3,1,nothing]"

Qecsim.BasicModels.FiveQubitCode — Method

FiveQubitCode()

Construct 5-qubit [5,1,3] code as a BasicCode.

Qecsim.BasicModels.SteaneCode — Method

SteaneCode()

Construct Steane [7,1,3] code as a BasicCode.

GenericModels