Core

Qecsim

Qecsim — Module

Package for simulating quantum error correction using stabilizer codes.

Qecsim.QecsimError — Type

QecsimError(msg::String)

Indicates an internal (core or models) error.

Model

Qecsim.Model — Module

Abstract types and methods for codes, error models and decoders.

Model.StabilizerCode

Qecsim.Model.StabilizerCode — Type

Abstract supertype for stabilizer codes.

Qecsim.Model.label — Function

label(code::StabilizerCode) -> String

Return a label suitable for use in plots and for grouping results.

Abstract method

This method should be implemented for concrete subtypes.

Qecsim.Model.logical_xs — Function

logical_xs(code::StabilizerCode) -> BitMatrix

Return the logical X operators in binary symplectic form.

Each row is an operator. The order should match that of logical_zs.

Abstract method

This method should be implemented for concrete subtypes.

Qecsim.Model.logical_zs — Function

logical_zs(code::StabilizerCode) -> BitMatrix

Return the logical Z operators in binary symplectic form.

Each row is an operator. The order should match that of logical_xs.

Abstract method

This method should be implemented for concrete subtypes.

Qecsim.Model.logicals — Function

logicals(code::StabilizerCode) -> BitMatrix

Return the logical operators in binary symplectic form.

Each row is an operator. X operators are stacked above Z operators in the order given by logical_xs and logical_zs.

Qecsim.Model.nkd — Function

nkd(code::StabilizerCode) -> Tuple{Int, Int, Union{Int,Nothing}}

Return a descriptor in the format (n, k, d), where n is the number of physical qubits, k is the number of logical qubits, and d is the distance of the code (or nothing if unknown).

Abstract method

This method should be implemented for concrete subtypes.

Qecsim.Model.stabilizers — Function

stabilizers(code::StabilizerCode) -> BitMatrix

Return the stabilizers in binary symplectic form.

Each row is a stabilizer generator. An overcomplete set of generators can be included to simplify decoding.

Abstract method

This method should be implemented for concrete subtypes.

Qecsim.Model.validate — Function

validate(code::StabilizerCode)

Perform sanity checks.

If any of the following fail then a QecsimError is thrown:

$S \odot S^T = 0$

$S \odot L^T = 0$

$L \odot L^T = \Lambda$

where $S$ and $L$ are the code stabilizers and logicals, respectively, and $\odot$ and $\Lambda$ are defined in bsp.

Model.ErrorModel

PauliTools

Qecsim.PauliTools — Module

Tools for Pauli strings and binary symplectic vectors / matrices.

Qecsim.PauliTools.bsp — Method

bsp(A::Union{BitVector, BitMatrix}, B::Union{BitVector, BitMatrix})

Return the binary symplectic product of A with B, given in binary symplectic form.

The binary symplectic product $\odot$ is defined as $A \odot B \equiv A \Lambda B \bmod 2$ where $\Lambda = \left[\begin{smallmatrix} 0 & I \\ I & 0 \end{smallmatrix}\right]$.

Examples

julia> a = BitVector([1, 0, 0, 0]);  # XI

julia> b = BitVector([0, 0, 1, 0]);  # ZI

julia> bsp(a', b)
1

julia> stabilizers = BitMatrix(  # 5-qubit stabilizers
       [1 0 0 1 0 0 1 1 0 0     # XZZXI
        0 1 0 0 1 0 0 1 1 0     # IXZZX
        1 0 1 0 0 0 0 0 1 1     # XIXZZ
        0 1 0 1 0 1 0 0 0 1]);  # ZXIXZ

julia> error = BitVector([0, 0, 1, 1, 0, 0, 1, 0, 1, 0]);  # IZXYI

julia> bsp(stabilizers, error)
4-element Vector{Int64}:
 0
 1
 1
 0

Qecsim.PauliTools.to_bsf — Method

to_bsf(pauli::Union{String, Iterable of String})

Convert the Pauli string operator(s) to binary symplectic form.

A single Pauli string is converted to a vector. A collection of Pauli strings is converted to a matrix where each row corresponds to a Pauli.

Examples

julia> to_bsf("XIZIY")
10-element BitVector:
 1
 0
 0
 0
 1
 0
 0
 1
 0
 1

julia> to_bsf(["XIZIY", "IXZYI"])
2×10 BitMatrix:
 1  0  0  0  1  0  0  1  0  1
 0  1  0  1  0  0  0  1  1  0

Qecsim.PauliTools.to_pauli — Method

to_pauli(bsf::Union{BitVector, BitMatrix})

Convert binary symplectic form to Pauli string operator(s).

A vector is converted to a single Pauli string. A matrix is converted row-by-row to a collection of Pauli strings.

Examples

julia> to_pauli(BitVector([1, 0, 0, 0, 1, 0, 0, 1, 0, 1]))
"XIZIY"

julia> to_pauli(BitMatrix([1 0 0 0 1 0 0 1 0 1; 0 1 0 1 0 0 0 1 1 0]))
2-element Vector{String}:
 "XIZIY"
 "IXZYI"