Core

Qecsim

Qecsim — Module

Package for simulating quantum error correction using stabilizer codes.

Qecsim.QecsimError — Type

QecsimError <: Exception

QecsimError(msg)

Construct an exception indicating an internal (core or models) error.

App

Qecsim.App — Module

Functions to run quantum error correction simulations and merge output data.

Qecsim.App.qec_run_once — Function

qec_run_once(code::StabilizerCode, error_model::ErrorModel, decoder::Decoder,
    p::Float64, rng::AbstractRNG=GLOBAL_RNG)

Run a stabilizer code error-decode-recovery (ideal) simulation and return run data.

Model

Qecsim.Model — Module

Abstract types and methods for codes, error models and decoders.

Model.AbstractModel

Qecsim.Model.AbstractModel — Type

Abstract supertype for models.

Qecsim.Model.label — Function

label(code::AbstractModel) -> String

Return a label suitable for use in plots and for grouping results.

Abstract method

This method should be implemented for concrete subtypes of AbstractModel.

Model.StabilizerCode

Qecsim.Model.StabilizerCode — Type

StabilizerCode <: AbstractModel

Abstract supertype for stabilizer codes.

Qecsim.Model.logical_xs — Function

logical_xs(code::StabilizerCode) -> BitMatrix

Return the logical X operators in binary symplectic form.

Each row is an operator. The order should match that of logical_zs.

Abstract method

This method should be implemented for concrete subtypes of StabilizerCode.

Qecsim.Model.logical_zs — Function

logical_zs(code::StabilizerCode) -> BitMatrix

Return the logical Z operators in binary symplectic form.

Each row is an operator. The order should match that of logical_xs.

Abstract method

This method should be implemented for concrete subtypes of StabilizerCode.

Qecsim.Model.logicals — Function

logicals(code::StabilizerCode) -> BitMatrix

Return the logical operators in binary symplectic form.

Each row is an operator. X operators are stacked above Z operators in the order given by logical_xs and logical_zs.

Qecsim.Model.nkd — Function

nkd(code::StabilizerCode) -> Tuple{Int, Int, Union{Int, Missing}}

Return a descriptor in the format (n, k, d), where n is the number of physical qubits, k is the number of logical qubits, and d is the distance of the code (or missing if unknown).

Abstract method

This method should be implemented for concrete subtypes of StabilizerCode.

Qecsim.Model.stabilizers — Function

stabilizers(code::StabilizerCode) -> BitMatrix

Return the stabilizers in binary symplectic form.

Each row is a stabilizer generator. An overcomplete set of generators can be included to simplify decoding.

Abstract method

This method should be implemented for concrete subtypes of StabilizerCode.

Qecsim.Model.validate — Function

validate(code::StabilizerCode)

Perform sanity checks.

If any of the following fail then a QecsimError is thrown:

$S \odot S^T = 0$

$S \odot L^T = 0$

$L \odot L^T = \Lambda$

where $S$ and $L$ are the code stabilizers and logicals, respectively, and $\odot$ and $\Lambda$ are defined in bsp.

Model.ErrorModel

Qecsim.Model.ErrorModel — Type

ErrorModel <: AbstractModel

Abstract supertype for error models.

Qecsim.Model.generate — Function

generate(error_model::ErrorModel, code::StabilizerCode, p::Float64,
         [rng::AbstractRNG=GLOBAL_RNG]) -> BitVector

Generate a new error in binary symplectic form according to the error_model and code, where p is typically the probability of an error on a single qubit.

Abstract method

This method should be implemented for concrete subtypes of ErrorModel.

Qecsim.Model.probability_distribution — Function

probability_distribution(error_model::ErrorModel, p::Float64) -> NTuple{4, Real}

Return the single-qubit probability distribution amongst Pauli I, X, Y and Z, where p is the overall probability of an error on a single qubit.

Abstract method [optional]

This method is not invoked by any core modules. Since it is often useful for decoders, it is provided as a template and concrete subtypes of ErrorModel are encouraged to implement it when appropriate, particularly for IID error models.

Model.Decoder

Qecsim.Model.Decoder — Type

Decoder <: AbstractModel

Abstract supertype for decoders.

Qecsim.Model.DecodeResult — Type

DecodeResult(recovery::AbstractVector{Bool})

Construct a decoding result including the recovery operation.

Warning

In the future, this type will be extended to include success flags and more.

Qecsim.Model.decode — Function

decode(decoder::Decoder, code::StabilizerCode, syndrome::AbstractVector{Bool};
       kwargs...) -> DecodeResult

Resolve a recovery operation for the given code and syndrome, or evaluate the success of decoding, as encapsulated in the decode result.

The syndrome has length equal to the number of code stabilizers, and element values of 0 or 1 (false or true) indicate whether the corresponding stabilizer does or does not commute with the error, respectively.

Keyword parameters kwargs may be provided by the client with context values such as error_model, error_probability and error. Most implementations will ignore such parameters; however, if they are used, implementations should declare them explicitly and treat them as optional.

Abstract method

This method should be implemented for concrete subtypes of Decoder.

PauliTools

Qecsim.PauliTools — Module

Tools for Pauli strings and binary symplectic vectors / matrices.

Qecsim.PauliTools.bsp — Function

bsp(A::AbstractVecOrMat{Bool}, B::AbstractVecOrMat{Bool})
    -> Union{Bool, BitVector, BitMatrix}

Return the binary symplectic product of A with B, given in binary symplectic form.

The binary symplectic product $\odot$ is defined as $A \odot B \equiv A \Lambda B \bmod 2$ where $\Lambda = \left[\begin{smallmatrix} 0 & I \\ I & 0 \end{smallmatrix}\right]$.

Examples

julia> a = BitVector([1, 0, 0, 0]);  # XI

julia> b = BitVector([0, 0, 1, 0]);  # ZI

julia> bsp(a', b)
true

julia> stabilizers = BitMatrix(  # 5-qubit stabilizers
       [1 0 0 1 0 0 1 1 0 0     # XZZXI
        0 1 0 0 1 0 0 1 1 0     # IXZZX
        1 0 1 0 0 0 0 0 1 1     # XIXZZ
        0 1 0 1 0 1 0 0 0 1]);  # ZXIXZ

julia> error = BitVector([0, 0, 1, 1, 0, 0, 1, 0, 1, 0]);  # IZXYI

julia> bsp(stabilizers, error)
4-element BitVector:
 0
 1
 1
 0

Qecsim.PauliTools.pack — Function

pack(bsf::AbstractVector{Bool}) -> Tuple{String, Int}

Pack a binary vector into a concise representation, typically for log output. See also unpack.

Examples

julia> a = BitVector([1, 0, 1, 0, 1, 1]);  # XYZ

julia> b = pack(a)  # (hex_value, length)
("2b", 6)
julia> unpack(b) == a
true

Qecsim.PauliTools.to_bsf — Function

to_bsf(pauli::Union{AbstractString, AbstractVector{<:AbstractString}})

Convert the Pauli string operator(s) to binary symplectic form.

A single Pauli string is converted to a vector. A vector of Pauli strings is converted to a matrix where each row corresponds to a Pauli.

Examples

julia> to_bsf("XIZIY")
10-element BitVector:
 1
 0
 0
 0
 1
 0
 0
 1
 0
 1

julia> to_bsf(["XIZIY", "IXZYI"])
2×10 BitMatrix:
 1  0  0  0  1  0  0  1  0  1
 0  1  0  1  0  0  0  1  1  0

Qecsim.PauliTools.to_pauli — Function

to_pauli(bsf::AbstractVecOrMat{Bool})

Convert the binary symplectic form to Pauli string operator(s).

A vector is converted to a single Pauli string. A matrix is converted row-by-row to a collection of Pauli strings.

Examples

julia> to_pauli(BitVector([1, 0, 0, 0, 1, 0, 0, 1, 0, 1]))
"XIZIY"

julia> to_pauli(BitMatrix([1 0 0 0 1 0 0 1 0 1; 0 1 0 1 0 0 0 1 1 0]))
2-element Vector{String}:
 "XIZIY"
 "IXZYI"

Qecsim.PauliTools.unpack — Function

unpack(packed_bsf::Tuple{String, Int}) -> BitVector

Unpack a binary vector from a concise representation, typically from log output. See also pack.

Examples

julia> a = ("2b", 6);  # (hex_value, length)

julia> b = unpack(a)  # XYZ
6-element BitVector:
 1
 0
 1
 0
 1
 1
julia> pack(b) == a
true

Qecsim.PauliTools.weight — Function

weight(bsf::AbstractVecOrMat{Bool}) -> Int

Return the weight of the binary symplectic form.

Examples

julia> weight(BitVector([1, 0, 0, 0, 1, 0, 0, 1, 0, 1]))
3

julia> weight(BitMatrix([1 0 0 0 1 0 0 1 0 1; 1 1 1 1 1 0 0 0 0 0]))
8

weight(pauli::Union{AbstractString, AbstractVector{<:AbstractString}}) -> Int

Return the weight of the Pauli string operator(s).

Examples

julia> weight("XIZIY")
3

julia> weight(["XIZIY", "XXXXX"])
8