Core

Qecsim

Qecsim — Module

Package for simulating quantum error correction using stabilizer codes.

Qecsim.QecsimError — Type

QecsimError <: Exception

QecsimError(msg)

Construct an exception indicating an internal (core or models) error.

App

Qecsim.App — Module

Functions to run quantum error correction simulations and merge output data.

Qecsim.App.qec_run_once — Function

qec_run_once(code, error_model, decoder, p::Float64, rng::AbstractRNG=GLOBAL_RNG)
    -> RunResult

Execute a stabilizer code error-decode-recovery (ideal) simulation and return run result.

The parameters code, error_model and decoder should be concrete subtypes or duck-typed implementations of StabilizerCode, ErrorModel and Decoder, respectively.

A simulation follows the following algorithm:

$S ←$ stabilizers(code)

$L ←$ logicals(code)

$e ←$ generate(error_model, code, p, rng)

$y ← S ⊙ e$

decode_result $←$ decode(decoder, code, $y$; kwargs...)

$r ←$ decode_result.recovery

sanity check: $S \odot (r \oplus e) = 0$

logical_commutations $← L \odot (r \oplus e)$

success $← L \odot (r \oplus e) = 0$

The kwargs passed to decode include error_model, p and error; most decoders will ignore these parameters. The decode method returns a DecodeResult. If decode_result.success and/or decode_result.logical_commutations are specified, they override the values of success and logical_commutations, irrespective of whether decode_result.recovery is specified or not. The value decode_result.custom_values is passed through in the run result.

See also RunResult.

TODO: doctest

Qecsim.App.RunResult — Type

RunResult(success::Bool, error_weight::Int,
          logical_commutations::Union{Nothing,BitVector}
          custom_values::Union{Nothing,Vector{<:Real}})

Construct a run result as returned by qec_run_once.

TODO: doctest

Qecsim.App.qec_run — Function

qec_run(code, error_model, decoder, p::Float64, random_seed=nothing;
        max_runs::Union{Int,Nothing}=nothing, max_failures::Union{Int,Nothing}=nothing)
    -> Dict

Execute stabilizer code error-decode-recovery (ideal) simulations many times and return aggregated run data.

The parameters code, error_model and decoder should be concrete subtypes or duck-typed implementations of StabilizerCode, ErrorModel and Decoder, respectively.

TODO: complete doc

Model

Qecsim.Model — Module

Abstract types and methods for codes, error models and decoders.

Model.AbstractModel

Qecsim.Model.AbstractModel — Type

Abstract supertype for models.

Qecsim.Model.label — Function

label(model) -> String

Return a label suitable for use in plots and for grouping results.

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of AbstractModel.

Model.StabilizerCode

Qecsim.Model.StabilizerCode — Type

StabilizerCode <: AbstractModel

Abstract supertype for stabilizer codes.

Qecsim.Model.logical_xs — Function

logical_xs(code) -> BitMatrix

Return the logical X operators in binary symplectic form.

Each row is an operator. The order should match that of logical_zs.

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of StabilizerCode.

Qecsim.Model.logical_zs — Function

logical_zs(code) -> BitMatrix

Return the logical Z operators in binary symplectic form.

Each row is an operator. The order should match that of logical_xs.

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of StabilizerCode.

Qecsim.Model.logicals — Function

logicals(code) -> BitMatrix

Return the logical operators in binary symplectic form.

Each row is an operator. X operators are stacked above Z operators in the order given by logical_xs and logical_zs.

Qecsim.Model.nkd — Function

nkd(code) -> Tuple{Int,Int,Union{Int,Missing}}

Return a descriptor in the format (n, k, d), where n is the number of physical qubits, k is the number of logical qubits, and d is the distance of the code (or missing if unknown).

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of StabilizerCode.

Qecsim.Model.stabilizers — Function

stabilizers(code) -> BitMatrix

Return the stabilizers in binary symplectic form.

Each row is a stabilizer generator. An overcomplete set of generators can be included to simplify decoding.

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of StabilizerCode.

Qecsim.Model.validate — Function

validate(code)

Perform sanity checks.

If any of the following fail then a QecsimError is thrown:

$S \odot S^T = 0$

$S \odot L^T = 0$

$L \odot L^T = \Lambda$

where $S$ and $L$ are the code stabilizers and logicals, respectively, and $\odot$ and $\Lambda$ are defined in PauliTools.bsp.

Model.ErrorModel

Qecsim.Model.ErrorModel — Type

ErrorModel <: AbstractModel

Abstract supertype for error models.

Qecsim.Model.generate — Function

generate(error_model, code, p::Float64, [rng::AbstractRNG=GLOBAL_RNG]) -> BitVector

Generate a new error in binary symplectic form according to the error_model and code, where p is typically the probability of an error on a single qubit.

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of ErrorModel.

Qecsim.Model.probability_distribution — Function

probability_distribution(error_model, p::Float64) -> NTuple{4,Real}

Return the single-qubit probability distribution amongst Pauli I, X, Y and Z, where p is the overall probability of an error on a single qubit.

Abstract method [optional]

This method is not invoked by any core modules. Since it is often useful for decoders, it is provided as a template and concrete subtypes or duck-typed implementations of ErrorModel are encouraged to implement it when appropriate, particularly for IID error models.

Model.Decoder

Qecsim.Model.Decoder — Type

Decoder <: AbstractModel

Abstract supertype for decoders.

Qecsim.Model.decode — Function

decode(decoder, code, syndrome::AbstractVector{Bool}; kwargs...) -> DecodeResult

Resolve a recovery operation for the given code and syndrome, or evaluate the success of decoding, as encapsulated in the decode result.

The syndrome has length equal to the number of code stabilizers, and element values of 0 or 1 (false or true) indicate whether the corresponding stabilizer does or does not commute with the error, respectively.

Keyword parameters kwargs may be provided by the client, e.g. App, with context values such as error_model, error_probability and error. Most implementations will ignore such parameters; however, if they are used, implementations should declare them explicitly and treat them as optional.

See also DecodeResult.

Abstract method

This method should be implemented for concrete subtypes or duck-typed implementations of Decoder.

Qecsim.Model.DecodeResult — Type

DecodeResult(success::Union{Nothing,Bool},
             recovery::Union{Nothing,AbstractVector{Bool}},
             logical_commutations::Union{Nothing,AbstractVector{Bool}}
             custom_values::Union{Nothing,AbstractVector{<:Real}})

DecodeResult(; success::Union{Nothing,Bool}=nothing,
             recovery::Union{Nothing,AbstractVector{Bool}}=nothing,
             logical_commutations::Union{Nothing,AbstractVector{Bool}}=nothing
             custom_values::Union{Nothing,AbstractVector{<:Real}}=nothing)

Construct a decoding result as returned by decode.

Typically decoders will provide a recovery operation and delegate the evaluation of success and logical_commutations to the client, e.g. App. Optionally, success and/or logical_commutations may be provided as overrides. At least one of recovery or success must be specified to allow a success value to be resolved. Additionally custom_values may be specified. If logical_commutations and/or custom_values are provided then they should be of consistent type and size over identically parameterized simulation runs so that they can be summed across runs.

See also decode.

Examples

julia> DecodeResult(; recovery=BitVector([1, 0, 1, 0, 1, 1]))  # typical use-case
DecodeResult{Nothing}(nothing, Bool[1, 0, 1, 0, 1, 1], nothing, nothing)

julia> DecodeResult(; success=true)  # override success
DecodeResult{Nothing}(true, nothing, nothing, nothing)

julia> DecodeResult(; success=false, logical_commutations=BitVector([1, 0]))  # override all
DecodeResult{Nothing}(false, nothing, Bool[1, 0], nothing)

julia> DecodeResult(; success=true, custom_values=[2.3, 4.1])  # custom values
DecodeResult{Vector{Float64}}(true, nothing, nothing, [2.3, 4.1])

julia> DecodeResult(true, nothing, nothing, [2.3, 4.1])  # positional parameters
DecodeResult{Vector{Float64}}(true, nothing, nothing, [2.3, 4.1])

julia> DecodeResult(; success=nothing, recovery=nothing)  # too few specified parameters
ERROR: QecsimError: at least one of 'success' or 'recovery' must be specified
Stacktrace:
[...]

PauliTools

Qecsim.PauliTools — Module

Tools for Pauli strings and binary symplectic vectors / matrices.

Qecsim.PauliTools.bsp — Function

bsp(A::AbstractVecOrMat{Bool}, B::AbstractVecOrMat{Bool})
    -> Union{Bool,BitVector,BitMatrix}

Return the binary symplectic product of A with B, given in binary symplectic form.

The binary symplectic product $\odot$ is defined as $A \odot B \equiv A \Lambda B \bmod 2$ where $\Lambda = \left[\begin{smallmatrix} 0 & I \\ I & 0 \end{smallmatrix}\right]$.

Examples

julia> a = BitVector([1, 0, 0, 0]);  # XI

julia> b = BitVector([0, 0, 1, 0]);  # ZI

julia> bsp(a', b)
true

julia> stabilizers = BitMatrix(  # 5-qubit stabilizers
       [1 0 0 1 0 0 1 1 0 0     # XZZXI
        0 1 0 0 1 0 0 1 1 0     # IXZZX
        1 0 1 0 0 0 0 0 1 1     # XIXZZ
        0 1 0 1 0 1 0 0 0 1]);  # ZXIXZ

julia> error = BitVector([0, 0, 1, 1, 0, 0, 1, 0, 1, 0]);  # IZXYI

julia> bsp(stabilizers, error)
4-element BitVector:
 0
 1
 1
 0

Qecsim.PauliTools.pack — Function

pack(bsf::AbstractVector{Bool}) -> Tuple{String,Int}

Pack a binary vector into a concise representation, typically for log output. See also unpack.

Examples

julia> a = BitVector([1, 0, 1, 0, 1, 1]);  # XZY

julia> b = pack(a)  # (hex_value, length)
("2b", 6)
julia> unpack(b) == a
true

Qecsim.PauliTools.to_bsf — Function

to_bsf(pauli::Union{AbstractString,AbstractVector{<:AbstractString}})
    -> Union{BitVector,BitMatrix}

Convert the Pauli string operator(s) to binary symplectic form.

A single Pauli string is converted to a vector. A vector of Pauli strings is converted to a matrix where each row corresponds to a Pauli.

Examples

julia> to_bsf("XIZIY")
10-element BitVector:
 1
 0
 0
 0
 1
 0
 0
 1
 0
 1

julia> to_bsf(["XIZIY", "IXZYI"])
2×10 BitMatrix:
 1  0  0  0  1  0  0  1  0  1
 0  1  0  1  0  0  0  1  1  0

Qecsim.PauliTools.to_pauli — Function

to_pauli(bsf::AbstractVecOrMat{Bool}) -> Union{String,Vector{String}}

Convert the binary symplectic form to Pauli string operator(s).

A vector is converted to a single Pauli string. A matrix is converted row-by-row to a collection of Pauli strings.

Examples

julia> to_pauli(BitVector([1, 0, 0, 0, 1, 0, 0, 1, 0, 1]))
"XIZIY"

julia> to_pauli(BitMatrix([1 0 0 0 1 0 0 1 0 1; 0 1 0 1 0 0 0 1 1 0]))
2-element Vector{String}:
 "XIZIY"
 "IXZYI"

Qecsim.PauliTools.unpack — Function

unpack(packed_bsf::Tuple{String,Int}) -> BitVector

Unpack a binary vector from a concise representation, typically from log output. See also pack.

Examples

julia> a = ("2b", 6);  # (hex_value, length)

julia> b = unpack(a)  # XZY
6-element BitVector:
 1
 0
 1
 0
 1
 1
julia> pack(b) == a
true

Qecsim.PauliTools.weight — Function

weight(bsf::AbstractVecOrMat{Bool}) -> Int

Return the weight of the binary symplectic form.

Examples

julia> weight(BitVector([1, 0, 0, 0, 1, 0, 0, 1, 0, 1]))
3

julia> weight(BitMatrix([1 0 0 0 1 0 0 1 0 1; 1 1 1 1 1 0 0 0 0 0]))
8

weight(pauli::Union{AbstractString,AbstractVector{<:AbstractString}}) -> Int

Return the weight of the Pauli string operator(s).

Examples

julia> weight("XIZIY")
3

julia> weight(["XIZIY", "XXXXX"])
8