Models

BasicModels

Qecsim.BasicModels — Module

Basic stabilizer codes

Qecsim.BasicModels.BasicCode — Type

BasicCode <: StabilizerCode

Basic code defined by its stabilizers and logical operators.

Qecsim.BasicModels.FiveQubitCode — Function

FiveQubitCode()

Construct 5-qubit [5,1,3] code as a BasicCode.

Qecsim.BasicModels.SteaneCode — Function

SteaneCode()

Construct Steane [7,1,3] code as a BasicCode.

GenericModels

Qecsim.GenericModels — Module

Generic error models and decoders compatible with any stabilizer codes.

Qecsim.GenericModels.SimpleErrorModel — Type

SimpleErrorModel <: ErrorModel

Abstract supertype for simple IID error models that generate errors based on the number of qubits and a probability distribution.

Qecsim.Model.generate — Method

generate(error_model::SimpleErrorModel, code::StabilizerCode, p::Float64,
         [rng::AbstractRNG=GLOBAL_RNG]) -> BitVector

Generate a new IID error based on probability_distribution. See also Model.generate.

Note

The method probability_distribution should be implemented for concrete subtypes of SimpleErrorModel.